Add inspection.ts
