+++ packages/core/src/index.ts
@@ -15,13 +15,13 @@ import { StateNode } from './StateNode.ts';
 // TODO: decide from where those should be exported
 export { and, not, or, stateIn } from './guards.ts';
 export { setup } from './setup.ts';
+export type { ActorSystem } from './system.ts';
 export type {
-  ActorSystem,
   InspectedActorEvent,
   InspectedEventEvent,
   InspectedSnapshotEvent,
   InspectionEvent
-} from './system.ts';
+} from './inspection.ts';
 export { toPromise } from './toPromise.ts';
 export {
   getAllOwnEventDescriptors as __unsafe_getAllOwnEventDescriptors,
